move issues to blog
