+++ pages/blog.tsx
@@ -1,4 +1,5 @@
 import type { NextPage } from 'next'
+import Link from 'next/link'
 import { listBlogposts } from 'utils/fetchIssues'
 import BaseReadingLayout from 'layouts/BaseReadingLayout'
 
@@ -8,7 +9,11 @@ const IssueList: NextPage<{ issueList: string[] }> = props => {
       <h1>All of my writing</h1>
       <ul>
         {props.issueList.map(issueSlug => (
-          <li key={issueSlug}>{issueSlug}</li>
+          <li key={issueSlug}>
+            <Link href={`issues/${issueSlug}`}>
+              <a>{issueSlug}</a>
+            </Link>
+          </li>
         ))}
       </ul>
     </BaseReadingLayout>
